requirements.txt
streamlit
tensorflow-cpu==2.11.0
numpy
Pillow
matplotlib
